I broke it last night but we're good now
I tried to fix the visual studio warning it always gives me about a system files but then I had 327 errors. The game ran fine but I reverted it to the previous commit. That didn't do anything so I reverted the revert then deleted the library. Once it rebuilt the library, everything was back to normal. Now I'm gonna do the encapsulation bit