11:17 am progress
Pause broke again but I turned it into a singleton and now it's good. Game over screen still won't turn off when I go to the menu from the game over button. But it's fine when I use the menu button from the pause menu so I don't know what the deal is. Button select broke too so once I figure out the game over screen I'll need to deal with that.