Added basic high score UI
No actual high score list yet but the menu seems to work. All buttons work as they should and the player can access the high score menu from the main menu. I made this way harder than it should be. I should have added the high score screen as the last child of the canvas so I wouldn't have had to change all the references to children in the button controller, game manager, pause manager, etc. Hopefully unity gets a better system for dealing with UI in the future because this fucking sucks